feat: move quiz editing to modal
